Add missing protection for nullptr in cut array
In case only cuts on the AOD filter bit are applied, and
no additional track cuts, the cut array typically is a
nullptr, leading to a segfault dereferencing the array.
The function AddTrackCuts is only called in case the
array is not null. Furthermore the function checks itself
that the array is not null.