Interface for generic BDT models in AliPhysics
This commit, prepared with @pfecchio, introduces an interface to the
treelite library. Treelite takes serialised models created by xgboost,
lightgbm and scikit-learn and turns them in C code.
The C code can then be compiled into a library that is used for
inference.
For the time being the interface allows the user to load a serialised
model from LightGBM, xgboost or a compiled library generated locally.
Local tests on a small dataset show that the inference with this
interface and with xgboost (in python) gives exactly the same results.