Add JE train helpers for LEGO train configuration
This allows for the extraction of AliEn production variables in the
LEGO train global variables. Control statements such as for loops are
not allowed in the global variables, so this function allows more
complicated extraction and assignment of variables.

Note that a class with static methods is used instead of a namespace
because the are easier to use with rootmaps.