Add option to set a user-configured event cuts object
In order to allow for user-configured event cuts
objects (in the AddTask) an option to set the
event cut object is added.

AliEventCuts in AliAnalysisTaskEmcal not
created dynamically any more but put on the
stack - simplifying the customization of the
object.

Access to the EventCuts object from outside
is possible via the function GetEventCuts.

fUseInternalEventSelection renamed to
fUseBuiltinEventSelection in order to
avoid shadowing in the EMCAL correction
task.

Deprecate function SetUseInternalEventSelection
due to shadowing in AliAnalysisTaskEmbeddingHelper

Fix additional dependencies raised by making the
object a stack attribute