get year from title and decide if construct shield
Implements ParseYear()

If the regular expression "year=YYYY" is found in title, then the year is used to decide whether to construct the ADA PMT shielding or not.
if the regex is not found, or the constructor without name-title is used, the ADA PMT shielding is constructed by default. 
In the later case you can still use the method  EnablePmtShieldingADA(const Bool_t flag) to alter this behavior.