Fix crash when recycling embedded events
Regardless of the builtin event selection setting, the embedding needs to be
able to track when embedded events are recycled. Traditionally, this was
handled in the event rejection hist, but it isn't defined when using the
new AliEventCuts based event selection.

To address this, we always define the event rejection hist and then:

- When using AliEventCuts: Define only the "RecycleEmbeddedEvents" bin
  in the event rejection histogram.
- When using the builtin event selection: Define all of the bins and then
  put the "RecycleEmbeddedEvents" bin as the last bin in the event
  rejection histogram.