Store internal event phys sel in separate variable
This allows a separation of the selected collision candidates and the
internal event selection. It's of course preferable if they are the
same, but sometimes this isn't trivially possible. In such cases, we can
now specify a narrower physics selection. Plus, it breaks users'
expectations when we silently overrode the physics selection specified
via `SelectCollisionCandidates()`.