Fully move AliEmcalJetTask AddTask to static method
Previously, we had two copies (one as a static method in the class, and one
as a .C macro), which then diverged when one got updated. This now
adopts the same approach as used in a number of other tasks, where the
macro just wraps the static method defined in the class.