Early execution break when recycling an external event
This takes advantage of a new feature in `AliAnalysisManager`. This way,
the users don't need to enable embedded event recycling, which should
reduced mistakes in the configuration. Note that this does slightly
change the procedure because the later user task internal event
counting won't be accurate - the user task won't be executed. That means
that the number of accepted events will still be the same, but if the
number of rejected events is required, then it must be retrieved from
the embedding helper. The number of accepted events is the value that
matters for the pt hard scaling, so everything should be exactly the
same from the users' perspective.