Revise TRD digits filtering
The classes to filter a small subset of the TRD digits during 
reconstruction for later studies of electron PID and high-pT tracks have 
been revised in preparation for the Pb-Pb production.

The AliTRDdigitsFilter class now inherits from the AliTRDdigitsTask 
class. The selection criteria were made more flexible, and selecting a 
small section of events based on centrality is supported.

The selection criteria still need some tuning.