Add jet matching to jet-h performance
This isn't ideal, but it's needed to embed in 15o. Otherwise, the LEGO
train has awful efficiency due to jobs crashing. It's only when
embedding full jets in that particular dataset, and investigations
haven't revealed a clear cause. However, putting the matching in the
user task is known to work, so that's the approach that we take here.

See the code for further details