AliAnalysisAlien: allow gridless MergeInfo and MergeOutput (#1034)
Due to the migration to JAlien one now needs to go through a TGrid
factory method in order to get a concrete instance of TGridCollection.
This fails when we do not have the TGrid available like in the case of
the release validation. For those cases, we therefore fix it by exposing
the actual AliEN plugin implementation to AliRoot and we create the
objects ourselves in the relevant places.