Automatic configuration of the energy smearing
Users no longer need to configure a smear model
by hand if they want to use a smeared patch energy
calculated from the FEE energy with the same
resolution as in data. The smearing is automatically
activated in data and MC. Currently the default
smearing is used - custom smear models would still
work.