DRAFT: Reject AOD events for which the relabeling of photon candidates fails (DATA & MC)
This change impacts only scenarios where the relabeling of AOD photon
candidates is performed within the AliV0Reader and NOT within the actual
AliAnalysisTask.

General info: After an event has passed the V0Reader's event selection,
the photon candidates are getting imported from the delta AOD file and
relabeled. This means that for every photon (skipping the candidate from now on)
from the delta AOD file its daughter tracks are tried to match with
two tracks from the set of all tracks within the event and relabeled on
success.

Before this change: If for a photon the relabeling fails (i.e. that
 the photon's daughter tracks can't be found within the event), then
its daughter tracks are set to negative and the photon will create
an entry under 'no tracks' in the
GammaConvV1_...->ConvCuts_...->IsPhotonSelected histogram and there
will be an error message. The event will not get rejected. The log
can get big since the error is shown for each photon where it fails.

With this change: If for a photon the relabeling fails, it will create
an entry in the same histogram as above. In addition, the quality of the
event will be set to 2 such that all AnalysisTasks which check the event
quality (I guess all) will reject this event. This will become visible
in the GammaConvV1_...->Cut Number...->...ESD Histograms->NEvents
histogram under 'Miss MC or inc. ev.' An error is shown at maximum once
per event (even less since it would be identical AliErrors)
Moreover, the filename of the current AOD file be added to the tree
of broken files. (Only for the GammaConvV1 task with this change)

The problem with failing relabeling was observed at GA_PbPb_AOD_MC
trainrun 1544, file /alice/sim/2019/LHC19h2a/297481/AOD/027/
also see: https://alice.its.cern.ch/jira/browse/ALIROOT-8304