PWGGA/GammaConv: refactor ProcessPhotonCandidates() take2
This change reduces the complexity of AliAnalysisTask_GammaConvV1::
ProcessPhotonCandidates significantly. In addition, the semantics
of the sharedElectron and toCloseV0s cuts are slightly changed
and the computational costs reduced by a factor > 2.

semantic changes:
sharedElectronCut():
A use the chi2 values of photons to decide which photon to reject
  (before, the choice was random)
B a photon that has been rejected, can not make other photons get
  rejected anymore. However, photons that have already been
  rejected because of it, will stay rejected. Note that as a result
  the order in which the photons get compared can affect the
  set of rejected photons. Finding the minimal set of photons that
  need to be rejected would be highly non-trivial and only sometimes
  make a small difference.
  (before, an already rejected photon could still kick out further photons)

toCloseV0sCut():
same as B above

B affects typically 10% of photons for the shared electron
cut and sometimes one out of 200 photons for the tooclose cut.

The reduced computational costs come from:
- consider each pair of photons only once (for both mentioned cuts)
  (before: twice)
- in case of added particles, determine for each photon only once
  if it originates from an accepted injector.
  (before: up to three times)