[PWGJE-23] Add singleton handling access to PYTHIA cross section file
In order to minimize the amount of calls to
the pyxsec(_hists).root files a singleton is
used to cache the cross section value and
just pass it instead of load it again, if the file
is the same.

The reading of the pythia cross section via
singleton class is implemented in
AliAanlysisTaskEmcal as second fast option,
default is kept with the standard option. The
fast option can be selected on request and
will replace the default method after a
transition period.