Use mask handler in trigger maker to read OCDB trigger mask
Use mask handler for trigger mask reading from the
OCDB and remove builting trigger mask reading. With
this the same code can be used for visualization of the
trigger mask and utilization in the trigger maker, so
the trigger maker can make use of a well tested and
unit-testable code.