Fix double counting of mixed triggers
Extra loop over the mixed triggers (OR of EMCLA or
DCAL trigger) became obsolate once the downscale
factors were added to the standard cache and
therefore treated as normal trigger class, however the
extra loop was not removed, leading to a double
counting of the trigger class when filling the
luminosity counting. Extra loop is removed in order to
prevent the double counting - mixed triggers will
always be treated as standard trigger class.