PWGGA/GammaConv: Various changes to in-Jet analysis
- Outsource determination if particle is in Jet into dedicated function
in MesonCuts
- This change also affects that the same pi0 cannot be filled multiple
times into the same histogram (that was the case before if the same pi0
is inside multiple jet cones). Now its always filled for the closest
jet.
- Changed definition when a true pi0 is in a jet (before it was in jet
if it is in a true jet, now its in a reconstructed jet). This definition
is closer to data and therefore one can now really check how many pi0
in-Jet candidates are real pi0s.
- Added InitJet function to initialize all jet vectors just once per
event
- Added option for rotation background to GammaConv task