PWGGA/GammaConv Fix for unsafe pointer issue
Up to now pointers to the aod mc track array were obtained
only once for the first event of an instance of the
analysis task. Since it is not guaranteed that the address of
this track array stays constant over the lifetime of an analysis task
this can produce unexpected behaviour.
With this change the pointer is reset for every event.