PWGGA/GammaConv: Change logic for minimum M02 cut in case of energy dependent cut
- Previously, the flag fUseM02 = 2 was set if an energy dependent cut on
M02 should be performed. In that case however, the cut value for the
lower M02 cut was overwritten and set to 0.1 per default (which is okay
as 0.1 is the standard). Since that is a bit confusing, the logic was
changed and now the lower M02 cut can be set via the cut value 6, 7, 8
and 9, while the actual values are then set in the function
CalculateMinM02.
- Changed the cuts in the add task accordingly

vla