[RUN3, PWGGA] Fix circle dependency
The `AliAnalysisTaskAO2DConverter` needed to know of `AliV0ReaderV1` and vice versa.
This commit adds the class `AliAODRelabelInterface` which is a new base class that inherits itself from the AliAnalysisTaskSE class and that AliV0ReaderV1 and AliAnalysisTaskAO2DConverter inherit from. There is also a .cxx file for this class which only exists to trigger the building of the lib so that it can be added inside the RUN3 and PWGG/GammaConvBase CMakeLists.txt.
This way both classes only need the one new class and do not need to know from each other.
Inside PWGGA/GammaConvBase/AliConversionPhotonCuts both classes were also needed because there in the function `GetTrack` one needs to check if the V0s are already relabeled.
PWGCF/FEMTOSCOPY/FemtoDream/ also depends on the V0Reader so I neede to change the CMakeList.txt to include the new header path for PWGGA/Common